C-x C-g now runs a simplified version of magit-grep.
C-x C-g was previously bound to ack-and-a-half, but it turns out that in
many cases (such as very large code bases), magit-grep is
infinity-bajillion times faster. So I bound C-x C-g to a new function
git-grep-dwim, which searches all git-tracked files in the current
project.

Note that if you want to search files not checked in to git, like log
files, you'll have to use something else. ack-and-a-half is now bound to
C-x C-a.

Also, deleted an old unused version of clojure-refactoring mode.