Don't "require" in defuns/vendor function.
The "vendor" function that we use as a catch-all code-loader was always
calling "require", which seems to me to defeat the purpose of having
autoloads. Elpa packages all have autoloads defined, and we should try
to take advantage of that.

In theory, this should reduce our start-up time, but so far it doesn't
seem to have a big effect (maybe down to 7 seconds, from 10). I think
that's because we still have quite a few explicit calls to require.